Fix slot_mapping dtype in NWOR fallback path
The fallback path calls reshape_and_cache_flash which expects int64
slot_mapping, but we were converting to int32. This caused a runtime
error: "expected scalar type Long but found Int"

Change: Convert slot_mapping to int64 in _fallback_commit instead of int32

+++ vllm/v1/nwor/draft_manager.py
@@ -235,8 +235,9 @@ class DraftCommitManager:
             value_accepted = entry._value_ref[accepted_indices]
             slot_accepted = entry._slot_ref[accepted_indices]
 
-            if slot_accepted.dtype != torch.int32:
-                slot_accepted = slot_accepted.to(dtype=torch.int32)
+            # reshape_and_cache_flash expects int64
+            if slot_accepted.dtype != torch.int64:
+                slot_accepted = slot_accepted.to(dtype=torch.int64)
 
             # Scale slicing ONLY in fallback
             k_scale = None
